Rename module to param
